Move "include/overlay.h" into "src/mkit/as/" to get it out of "include/".
